Break out internal solvers into own components
