+++ drake/solvers/CMakeLists.txt
@@ -36,6 +36,7 @@ list(APPEND optimization_files
   equality_constrained_qp_solver.cc
   linear_system_solver.cc
   mathematical_program.cc
+  mathematical_program_api.cc
   mathematical_program_solver_interface.cc
   moby_lcp_solver.cc
   no_dreal.cc
