+++ drake/solvers/mathematical_program.cc
@@ -27,6 +27,9 @@
 #include "drake/solvers/snopt_solver.h"
 #include "drake/solvers/symbolic_extraction.h"
 
+// Note that the file mathematical_program_api.cc also contains some of the
+// implementation of mathematical_program.h
+
 namespace drake {
 namespace solvers {
 
@@ -589,39 +592,15 @@ MathematicalProgram::AddLinearMatrixInequalityConstraint(
   return AddConstraint(constraint, vars);
 }
 
-int MathematicalProgram::FindDecisionVariableIndex(const Variable& var) const {
-  auto it = decision_variable_index_.find(var.get_id());
-  if (it == decision_variable_index_.end()) {
-    ostringstream oss;
-    oss << var << " is not a decision variable in the mathematical program, "
-                  "when calling GetSolution.\n";
-    throw runtime_error(oss.str());
-  }
-  return it->second;
-}
+// Note that FindDecisionVariableIndex is implemented in
+// mathematical_program_api.cc instead of this file.
 
 double MathematicalProgram::GetSolution(const Variable& var) const {
   return x_values_[FindDecisionVariableIndex(var)];
 }
 
-void MathematicalProgram::SetDecisionVariableValues(
-    const Eigen::Ref<const Eigen::VectorXd>& values) {
-  SetDecisionVariableValues(decision_variables_, values);
-}
-
-void MathematicalProgram::SetDecisionVariableValues(
-    const Eigen::Ref<const VectorXDecisionVariable>& variables,
-    const Eigen::Ref<const Eigen::VectorXd>& values) {
-  DRAKE_ASSERT(values.rows() == variables.rows());
-  for (int i = 0; i < values.rows(); ++i) {
-    x_values_[FindDecisionVariableIndex(variables(i))] = values(i);
-  }
-}
-
-void MathematicalProgram::SetDecisionVariableValue(const Variable& var,
-                                                   double value) {
-  x_values_[FindDecisionVariableIndex(var)] = value;
-}
+// Note that SetDecisionVariableValue and SetDecisionVariableValues are
+// implemented in mathematical_program_api.cc instead of this file.
 
 SolutionResult MathematicalProgram::Solve() {
   // This implementation is simply copypasta for now; in the future we will
